Updated dummy API
Added enum of Binus' Locations
Changed position's properties to use Double instead of int
Updated dummy service

+++ src/main/java/com/binus/indoornavigation/model/Position.java
@@ -9,7 +9,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class Position {
 
-    private int latitude;
-    private int longitude;
+    private Double latitude;
+    private Double longitude;
 
 }
